Fixed an issue which caused failing `sys.snapshot` queries if the `data.path` of an existing fs repository was not configured anymore.
The problem was that crate was adding repositories 
to its own SysRepositoryService without verifying them, 
while Elasticsearch verifies the repositories when 
they are read form ClusterState.

This lead to a state that crates service had repositories
which were ignored by elasticsearch. When crate
was fetching the snapshots from this repositories
from elasticsearch the requests failed.

To avoid this problem i exposed the repositories
in the Elasticsearch service and use it directly.
Therefore i could also delete the crate service.